shift applicationStatus to same file as model
