--- src/constant/applicationStatus.ts
@@ -1,7 +1,0 @@
-export enum ApplicationStatus {
-  ACCEPTED = 'accepted',
-  OFFERED = 'offered',
-  PENDING = 'pending',
-  INTERVIEWED = 'interviewed',
-  REJECTED = 'rejected',
-}
